agents/pyproject.toml
[project]
name = "arcane-auditor-agents"
version = "0.1.0"
description = "Autonomous agent wrapper for Arcane Auditor deterministic code review"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.16.1",
    "pydantic>=2.0",
    "pygithub>=2.1.1",
    "pyyaml>=6.0",
]

[project.scripts]
arcane-agent = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "fix_templates"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]
